+++ src/renderer/components/markdown/TheMarkdown.vue
@@ -1,7 +1,7 @@
 <template>
   <div class="markdown markdown-body">
     <PerfectScrollbar>
-      <div v-html="renderer" />
+      <div v-html="renderedHtml" />
     </PerfectScrollbar>
   </div>
 </template>
@@ -11,11 +11,14 @@ import { useAppStore } from '@/store/app'
 import { useSnippetStore } from '@/store/snippets'
 import sanitizeHtml from 'sanitize-html'
 import hljs from 'highlight.js'
-import { computed, onBeforeUnmount, onMounted, ref, watch } from 'vue'
+import { computed, onBeforeUnmount, onMounted, ref, watch, nextTick } from 'vue'
 import { ipc, store } from '@/electron'
 import { marked } from 'marked'
 import mermaid from 'mermaid'
-import { useHljsTheme } from '@/composable'
+import { useHljsTheme, goToSnippet } from '@/composable'
+import { useCodemirror } from '@/composable/codemirror'
+import { nanoid } from 'nanoid'
+import { useMagicKeys } from '@vueuse/core'
 
 const isDev = import.meta.env.DEV
 
@@ -24,17 +27,29 @@ interface Props {
   scale?: number
 }
 
+interface Editors {
+  id: string
+  value: string
+  lang: string
+}
+
 const props = withDefaults(defineProps<Props>(), {
   scale: 1
 })
 
 const appStore = useAppStore()
 const snippetStore = useSnippetStore()
+const { escape } = useMagicKeys()
+
+const renderedHtml = ref()
+
+const editors: Editors[] = []
 
 const forceRefresh = ref()
 const preTagBg = computed(() =>
   appStore.isLightTheme ? '#fff' : 'var(--color-contrast-high)'
 )
+const fontFamily = computed(() => appStore.editor.fontFamily)
 
 const init = () => {
   const renderer: marked.RendererObject = {
@@ -42,14 +57,31 @@ const init = () => {
       if (lang === 'mermaid') {
         return `<div class="mermaid">${code}</div><br>`
       } else {
-        const language = hljs.getLanguage(lang) ? lang : 'plaintext'
-        return `<pre><code class="language-${lang}">${
-          hljs.highlight(code, { language }).value
-        }</code></pre>`
+        if (appStore.markdown.codeRenderer === 'highlight.js') {
+          const language = hljs.getLanguage(lang) ? lang : 'plaintext'
+          return `<pre><code class="language-${lang}">${
+            hljs.highlight(code, { language }).value
+          }</code></pre>`
+        } else {
+          const id = nanoid(6)
+
+          editors.push({
+            id,
+            value: code,
+            lang
+          })
+
+          return `<div id="${id}"></div>`
+        }
       }
     },
     link (href: string, title: string, text: string) {
-      return `<a href="${href}" class="external">${text}</a>`
+      if (/^masscode:\/\/snippets/.test(href)) {
+        const id = href.split('/').pop()
+        return `<a href="${href}" class="snippet-link" data-snippet-id="${id}">${text}</a>`
+      } else {
+        return `<a href="${href}" class="external">${text}</a>`
+      }
     }
   }
 
@@ -68,11 +100,14 @@ const initMermaid = () => {
 }
 
 onMounted(() => {
-  initMermaid()
+  render()
 })
 
-const getRenderer = () => {
+const render = () => {
+  if (!props.value) return
+
   const raw = marked.parse(props.value)
+
   let html = sanitizeHtml(raw, {
     allowedTags: [
       'h1',
@@ -152,9 +187,12 @@ const getRenderer = () => {
         'class',
         'type',
         'checked',
-        'disabled'
+        'disabled',
+        'id',
+        'data-*'
       ]
-    }
+    },
+    allowedSchemes: ['http', 'https', 'masscode']
   })
 
   const re = /src="\.\//g
@@ -164,16 +202,22 @@ const getRenderer = () => {
     ? html.replace(re, `src="file://${path}/`)
     : html.replace(re, `src="${path}/`)
 
-  return html
+  renderedHtml.value = html
+
+  nextTick(() => initMermaid())
 }
 
-const renderer = computed(() => getRenderer())
-
-const openExternal = (e: Event) => {
+const onLink = async (e: Event) => {
   const el = e.target as HTMLAnchorElement
   e.preventDefault()
+
   if (el.classList.contains('external')) {
     ipc.invoke('main:open-url', el.href)
+  }
+
+  if (el.classList.contains('snippet-link')) {
+    const { snippetId } = el.dataset
+    if (snippetId) goToSnippet(snippetId, true)
   }
 }
 
@@ -209,14 +253,36 @@ watch(
   { immediate: true }
 )
 
+watch(renderedHtml, () => {
+  nextTick(() => {
+    editors.forEach(i => {
+      useCodemirror(i.id, {
+        value: i.value,
+        mode: i.lang
+      })
+    })
+  })
+})
+
+watch(
+  () => props.value,
+  () => {
+    render()
+  }
+)
+
+watch(escape, () => {
+  snippetStore.isMarkdownPreview = false
+})
+
 init()
 
 onMounted(() => {
-  document.addEventListener('click', openExternal)
+  document.addEventListener('click', onLink)
 })
 
 onBeforeUnmount(() => {
-  document.removeEventListener('click', openExternal)
+  document.removeEventListener('click', onLink)
 })
 
 window.addEventListener('resize', () => {
@@ -237,6 +303,21 @@ window.addEventListener('resize', () => {
   }
   :deep(.ps) {
     height: v-bind(height);
+  }
+  :deep(.CodeMirror) {
+    height: 100%;
+    padding: var(--spacing-xs);
+    font-size: 0.85em;
+    font-family: v-bind(fontFamily);
+  }
+  :deep(.CodeMirror-line) {
+    padding: 0 var(--spacing-sm);
+    &:first-child {
+      padding-top: var(--spacing-xs);
+    }
+    &:last-child {
+      padding-bottom: var(--spacing-xs);
+    }
   }
 }
 </style>