+++ src/shared/types/main/analytics.d.ts
@@ -8,6 +8,7 @@ type SnippetEvents =
   | 'add-to-favorites'
   | 'code-preview'
   | 'copy'
+  | 'copy-link'
   | 'create-screenshot'
   | 'delete-fragment'
   | 'delete-from-favorites'
@@ -16,7 +17,7 @@ type SnippetEvents =
   | 'format'
   | 'markdown-preview'
   | 'move-to-trash'
-  | 'presentation-mode'
+  | 'restore-from-trash'
   | 'search'
   | 'set-language'
 
@@ -37,6 +38,7 @@ type AppEvents =
   | 'set-language'
 
 type ApiEvents = 'snippet-create'
+type PageEvents = 'main' | 'preferences' | 'presentation'
 
 type TrackSnippetEvents = CombineWith<SnippetEvents, 'snippets'>
 type TrackFolderEvents = CombineWith<FolderEvents, 'folders'>
@@ -50,5 +52,4 @@ export type TrackEvents =
   | TrackTagEvents
   | TrackAppEvents
   | TrackApiEvents
-  | 'main'
-  | 'preferences'
+  | PageEvents